refa: rename henrikfricke -> superluminar
