+++ apps/eureka/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.henrikfricke</groupId>
+        <groupId>com.superluminar</groupId>
         <artifactId>demo</artifactId>
         <version>1.0.0</version>
         <relativePath>../../pom.xml</relativePath> <!-- lookup parent from repository -->
