--- apps/eureka/src/main/java/com/henrickfricke/eureka/EurekaApplication.java
@@ -1,15 +1,0 @@
-package com.henrickfricke.eureka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class EurekaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EurekaApplication.class, args);
-	}
-
-}
