--- apps/eureka/src/test/java/com/henrickfricke/eureka/EurekaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.henrickfricke.eureka;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EurekaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
