--- apps/events/src/main/java/com/henrickfricke/events/EventsApplication.java
@@ -1,15 +1,0 @@
-package com.henrickfricke.events;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class EventsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EventsApplication.class, args);
-	}
-
-}
