--- apps/events/src/main/java/com/henrickfricke/events/controller/HelloWorldController.java
@@ -1,12 +1,0 @@
-package com.henrickfricke.events.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloWorldController {
-    @GetMapping("/")
-    public String index() {
-        return "Hello World :)";
-    }
-}
