--- apps/events/src/main/java/com/henrickfricke/events/dto/EventDTO.java
@@ -1,16 +1,0 @@
-package com.henrickfricke.events.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import javax.validation.constraints.NotEmpty;
-
-public class EventDTO {
-    @NotEmpty
-
-    public final String identifier;
-
-    EventDTO(@JsonProperty("identifier") String identifier) {
-        super();
-        this.identifier = identifier;
-    }
-}
