--- apps/events/src/main/java/com/henrickfricke/events/repository/EventsRepository.java
@@ -1,13 +1,0 @@
-package com.henrickfricke.events.repository;
-
-import com.henrickfricke.events.model.Event;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Date;
-import java.util.List;
-
-public interface EventsRepository extends JpaRepository<Event, Long> {
-    List<Event> findByPublishedAtAfter(Date publishedAt);
-
-    List<Event> findByIdentifier(String identifier);
-}
