+++ apps/events/src/main/java/com/superluminar/events/EventsApplication.java
@@ -1,0 +1,15 @@
+package com.superluminar.events;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class EventsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EventsApplication.class, args);
+	}
+
+}
