+++ apps/events/src/main/java/com/superluminar/events/controller/HelloWorldController.java
@@ -1,0 +1,12 @@
+package com.superluminar.events.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HelloWorldController {
+    @GetMapping("/")
+    public String index() {
+        return "Hello World :)";
+    }
+}
