--- apps/events/src/test/java/com/henrickfricke/events/EventsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.henrickfricke.events;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EventsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
