+++ apps/events/src/test/java/com/superluminar/events/EventsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.superluminar.events;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EventsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
