--- apps/tutorials/src/main/java/com/henrikfricke/tutorials/TutorialsApplication.java
@@ -1,15 +1,0 @@
-package com.henrikfricke.tutorials;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableFeignClients
-public class TutorialsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TutorialsApplication.class, args);
-	}
-
-}
