--- apps/tutorials/src/main/java/com/henrikfricke/tutorials/controller/StartpageController.java
@@ -1,16 +1,0 @@
-package com.henrikfricke.tutorials.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@Controller
-public class StartpageController {
-    @GetMapping("/")
-    public String index(@RequestParam(name = "name", required = false, defaultValue = "World") String name,
-            Model model) {
-        model.addAttribute("name", name);
-        return "startpage";
-    }
-}
