--- apps/tutorials/src/main/java/com/henrikfricke/tutorials/repository/TutorialRepository.java
@@ -1,11 +1,0 @@
-package com.henrikfricke.tutorials.repository;
-
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import com.henrikfricke.tutorials.model.Tutorial;
-
-public interface TutorialRepository extends JpaRepository<Tutorial, Long> {
-    List<Tutorial> findByPublished(boolean published);
-
-    List<Tutorial> findByTitleContaining(String title);
-}
