+++ apps/tutorials/src/main/java/com/superluminar/tutorials/TutorialsApplication.java
@@ -1,0 +1,15 @@
+package com.superluminar.tutorials;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients
+public class TutorialsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TutorialsApplication.class, args);
+	}
+
+}
