+++ apps/tutorials/src/main/java/com/superluminar/tutorials/controller/StartpageController.java
@@ -1,0 +1,16 @@
+package com.superluminar.tutorials.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@Controller
+public class StartpageController {
+    @GetMapping("/")
+    public String index(@RequestParam(name = "name", required = false, defaultValue = "World") String name,
+            Model model) {
+        model.addAttribute("name", name);
+        return "startpage";
+    }
+}
