+++ apps/tutorials/src/main/resources/application.yaml
@@ -25,6 +25,6 @@ logging:
   level:
     feign: INFO
     com:
-      henrikfricke:
+      superluminar:
         tutorials:
           client: INFO
