--- apps/tutorials/src/test/java/com/henrikfricke/tutorials/TutorialsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.henrikfricke.tutorials;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TutorialsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
