+++ apps/tutorials/src/test/java/com/superluminar/tutorials/TutorialsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.superluminar.tutorials;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TutorialsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
