+++ pom.xml
@@ -3,7 +3,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.henrikfricke</groupId>
+    <groupId>com.superluminar</groupId>
     <artifactId>demo</artifactId>
     <version>1.0.0</version>
     <packaging>pom</packaging>
